+++ src/test/test_num_spline.inl
@@ -71,9 +71,9 @@ std::string doubleToString(double iX, int iDigitsMantissa, int iDigitsExponent)
 	LASS_ASSERT(splitted.size() == 2);
 
 	std::string mantissa = splitted[0];
-	LASS_ASSERT(mantissa.length() == iDigitsMantissa + 2);
+	LASS_ASSERT(mantissa.length() == static_cast<size_t>(iDigitsMantissa + 2));
 	const std::string negativeZero = "-0." + std::string(iDigitsMantissa - 1, '0');
-	LASS_ASSERT(negativeZero.length() == iDigitsMantissa + 2);
+	LASS_ASSERT(negativeZero.length() == static_cast<size_t>(iDigitsMantissa + 2));
 	if (mantissa == negativeZero)
 	{
 		mantissa[0] = '+';
@@ -84,7 +84,7 @@ std::string doubleToString(double iX, int iDigitsMantissa, int iDigitsExponent)
 	buffer << std::setw(iDigitsExponent + 1) << std::showpos << std::internal 
 		<< std::setfill('0') << e;
 	std::string exponent = buffer.str();
-	LASS_ASSERT(exponent.length() == iDigitsExponent + 1);
+	LASS_ASSERT(exponent.length() == static_cast<size_t>(iDigitsExponent + 1));
 	if (exponent[0] == '0')
 	{
 		LASS_ASSERT(e == 0);
@@ -92,7 +92,7 @@ std::string doubleToString(double iX, int iDigitsMantissa, int iDigitsExponent)
 	}
 
 	const std::string result = mantissa + "e" + exponent;
-	LASS_ASSERT(result.length() == iDigitsMantissa + iDigitsExponent + 4);
+	LASS_ASSERT(result.length() == static_cast<size_t>(iDigitsMantissa + iDigitsExponent + 4));
 
 	return result;
 }
